+++ tests/Chat/ChatSmokeTests.cs
@@ -475,38 +475,4 @@ public partial class ChatSmokeTests : SyncAsyncTestBase
             Assert.That(additionalPropertyProperty.ValueKind, Is.EqualTo(JsonValueKind.True));
         }
     }
-
-    [Test]
-    [NonParallelizable]
-    public async Task HelloWorldChatWithTracingAndMetrics()
-    {
-        using var _ = TestAppContextSwitchHelper.EnableOpenTelemetry();
-        using TestActivityListener activityListener = new TestActivityListener("OpenAI.ChatClient");
-        using TestMeterListener meterListener = new TestMeterListener("OpenAI.ChatClient");
-
-        ChatClient client = GetTestClient<ChatClient>(TestScenario.Chat);
-        IEnumerable<ChatMessage> messages = [new UserChatMessage("Hello, world!")];
-        ClientResult<ChatCompletion> result = IsAsync
-            ? await client.CompleteChatAsync(messages)
-            : client.CompleteChat(messages);
-
-        Assert.AreEqual(1, activityListener.Activities.Count);
-        TestActivityListener.ValidateChatActivity(activityListener.Activities.Single(), result.Value);
-
-        List<TestMeasurement> durations = meterListener.GetMeasurements("gen_ai.client.operation.duration");
-        Assert.AreEqual(1, durations.Count);
-        ValidateChatMetricTags(durations.Single(), result.Value);
-
-        List<TestMeasurement> usages = meterListener.GetMeasurements("gen_ai.client.token.usage");
-        Assert.AreEqual(2, usages.Count);
-
-        Assert.True(usages[0].tags.TryGetValue("gen_ai.token.type", out var type));
-        Assert.IsInstanceOf<string>(type);
-
-        TestMeasurement input = (type is "input") ? usages[0] : usages[1];
-        TestMeasurement output = (type is "input") ? usages[1] : usages[0];
-
-        Assert.AreEqual(result.Value.Usage.InputTokens, input.value);
-        Assert.AreEqual(result.Value.Usage.OutputTokens, output.value);
-    }
 }
